requirements.txt
numpy
matplotlib
python-dotenv
scipy
requests